pyproject.toml
[project]
python = "^3.8"

[tool.poetry]
name = "gp-registrations-management-information"
version = "0.0.1-alpha"
description = "TODO"
authors = [
    #TODO add authors
]
readme = 'README.md'
license = "MIT"
repository = "https://github.com/NHSDigital/gp-registrations-management-information"
homepage = ""
keywords = ["healthcare", "uk", "nhs"] #TODO add additional keywords


[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
proxygen-cli = "^2.0.10"
pytest-nhsd-apim = "^3.3.0"

[tool.poetry.scripts]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(name)s.%(module)s.%(funcName)s - %(levelname)s - %(message)s"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s - %(name)s.%(module)s.%(funcName)s - %(levelname)s - %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
